implement signup email verification

+++ src/users/users.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Delete,
   Query,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -47,13 +48,34 @@ export class UsersController {
 
   @ApiNotFoundResponse({ description: 'not found' })
   @Patch(':id')
-  update(@Param('id') id: number, @Body() updateUserDto: UpdateUserDto) {
+  update(
+    @Param('id') id: number,
+    @Body() updateUserDto: UpdateUserDto,
+    @CurrentUser() user: User,
+  ) {
+    if (user.id !== +id) {
+      throw new ForbiddenException('only account owner can update the info');
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
   @ApiNotFoundResponse({ description: 'not found' })
   @Delete(':id')
-  remove(@Param('id') id: number) {
+  remove(@Param('id') id: number, @CurrentUser() user: User) {
+    if (user.id !== +id) {
+      throw new ForbiddenException('only account owner can delete it');
+    }
     return this.usersService.remove(id);
+  }
+
+  @ApiNotFoundResponse({ description: 'not found' })
+  @Get(':id/verify/:token')
+  verifyEmail(@Param('id') id: number, @Param('token') token: string) {
+    return this.usersService.verfiyEmail(id, token);
+  }
+
+  @Get(':id/send-email-verification-token')
+  sendEmailVerificationToken(@Param('id') id: number) {
+    return this.usersService.SendEmailVerificationToken(id);
   }
 }
